requirements.txt
torch
torchvision
torchaudio
torchmetrics
tensorflow_hub
spacy
wikipedia
pandas
tqdm
nltk
pillow
tables
wikipedia
transformers
datasets
sentence_transformers
matplotlib

terminaltables
# pycocotools
git+https://github.com/openai/CLIP.git
git+https://github.com/modestyachts/ImageNetV2_pytorch

img2pdf
tiktoken
open_clip_torch
wandb
openai
